auth-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic-settings==2.5.2
python-dotenv==1.0.1
bcrypt==4.0.1
email-validator==2.2.0
redis==5.1.1
